Simplify the refactor (thank you Andrei!)
