+++ packages/edit-navigation/src/components/menu-editor/index.js
@@ -13,28 +13,9 @@ import { useMemo } from '@wordpress/element';
  */
 import useMenuItems from './use-menu-items';
 import useNavigationBlocks from './use-navigation-blocks';
-import useCreateMissingMenuItems from './use-create-missing-menu-items';
-import useSaveNavigationBlocks from './use-save-navigation-blocks';
 import MenuEditorShortcuts from './shortcuts';
 import BlockEditorPanel from './block-editor-panel';
 import NavigationStructurePanel from './navigation-structure-panel';
-
-const useMenuEditor = ( menuId ) => {
-	const query = useMemo( () => ( { menus: menuId, per_page: -1 } ), [
-		menuId,
-	] );
-	const menuItems = useMenuItems( query );
-	const { blocks, setBlocks, menuItemsRef } = useNavigationBlocks(
-		menuItems
-	);
-	const saveBlocks = useSaveNavigationBlocks( blocks, menuItemsRef, query );
-	const { createMissingMenuItems, onCreated } = useCreateMissingMenuItems(
-		menuItemsRef
-	);
-	const eventuallySaveBlocks = () => onCreated( () => saveBlocks() );
-
-	return [ blocks, setBlocks, eventuallySaveBlocks, createMissingMenuItems ];
-};
 
 export default function MenuEditor( {
 	menuId,
@@ -42,12 +23,20 @@ export default function MenuEditor( {
 	onDeleteMenu,
 } ) {
 	const isLargeViewport = useViewportMatch( 'medium' );
-	const [
-		blocks,
-		setBlocks,
-		saveBlocks,
+	const query = useMemo( () => ( { menus: menuId, per_page: -1 } ), [
+		menuId,
+	] );
+	const {
+		menuItems,
+		saveMenuItems,
 		createMissingMenuItems,
-	] = useMenuEditor( menuId );
+		onCreated,
+	} = useMenuItems( query );
+	const { blocks, setBlocks, menuItemsRef } = useNavigationBlocks(
+		menuItems
+	);
+	const eventuallySaveItems = () =>
+		onCreated( () => saveMenuItems( blocks, menuItemsRef ) );
 
 	return (
 		<div className="edit-navigation-menu-editor">
@@ -58,7 +47,7 @@ export default function MenuEditor( {
 				value={ blocks }
 				onInput={ ( updatedBlocks ) => setBlocks( updatedBlocks ) }
 				onChange={ ( updatedBlocks ) => {
-					createMissingMenuItems( updatedBlocks );
+					createMissingMenuItems( updatedBlocks, menuItemsRef );
 					setBlocks( updatedBlocks );
 				} }
 				settings={ {
@@ -68,13 +57,13 @@ export default function MenuEditor( {
 				} }
 			>
 				<BlockEditorKeyboardShortcuts />
-				<MenuEditorShortcuts saveBlocks={ saveBlocks } />
+				<MenuEditorShortcuts saveBlocks={ eventuallySaveItems } />
 				<NavigationStructurePanel
 					blocks={ blocks }
 					initialOpen={ isLargeViewport }
 				/>
 				<BlockEditorPanel
-					saveBlocks={ saveBlocks }
+					saveBlocks={ eventuallySaveItems }
 					menuId={ menuId }
 					onDeleteMenu={ onDeleteMenu }
 				/>
