+++ packages/edit-navigation/src/components/menu-editor/use-create-missing-menu-items.js
@@ -16,41 +16,37 @@ import { flattenBlocks } from './helpers';
  * 1) limit the amount of requests processed at the same time
  * 2) save the menu only after all requests are finalized
  *
- * @param {Object} menuItemsRef Ref holding a mapping from clientId to it's related menuItem
  * @return {function(*=): void} Function registering it's argument to be called once all
  * 													    currently processed menuItems are created.
  */
-export default function useCreateMissingMenuItems( menuItemsRef ) {
+export default function useCreateMissingMenuItems() {
 	const promiseQueueRef = useRef( new PromiseQueue() );
 	const enqueuedBlocksIds = useRef( [] );
-	const createMissingMenuItems = useCallback(
-		( blocks ) => {
-			for ( const { clientId, name } of flattenBlocks( blocks ) ) {
-				// No need to create menuItems for the wrapping navigation block
-				if ( name === 'core/navigation' ) {
-					continue;
-				}
-				// Menu item was already created
-				if ( clientId in menuItemsRef.current ) {
-					continue;
-				}
-				// Already in the queue
-				if ( enqueuedBlocksIds.current.includes( clientId ) ) {
-					continue;
-				}
-				enqueuedBlocksIds.current.push( clientId );
-				promiseQueueRef.current.enqueue( () =>
-					createDraftMenuItem( clientId ).then( ( menuItem ) => {
-						menuItemsRef.current[ clientId ] = menuItem;
-						enqueuedBlocksIds.current.splice(
-							enqueuedBlocksIds.current.indexOf( clientId )
-						);
-					} )
-				);
+	const createMissingMenuItems = ( blocks, menuItemsRef ) => {
+		for ( const { clientId, name } of flattenBlocks( blocks ) ) {
+			// No need to create menuItems for the wrapping navigation block
+			if ( name === 'core/navigation' ) {
+				continue;
 			}
-		},
-		[ menuItemsRef.current ]
-	);
+			// Menu item was already created
+			if ( clientId in menuItemsRef.current ) {
+				continue;
+			}
+			// Already in the queue
+			if ( enqueuedBlocksIds.current.includes( clientId ) ) {
+				continue;
+			}
+			enqueuedBlocksIds.current.push( clientId );
+			promiseQueueRef.current.enqueue( () =>
+				createDraftMenuItem( clientId ).then( ( menuItem ) => {
+					menuItemsRef.current[ clientId ] = menuItem;
+					enqueuedBlocksIds.current.splice(
+						enqueuedBlocksIds.current.indexOf( clientId )
+					);
+				} )
+			);
+		}
+	};
 	const onCreated = useCallback(
 		( callback ) => promiseQueueRef.current.then( callback ),
 		[ promiseQueueRef.current ]
